test2

+++ src/modules/sensors/S8/S8.cpp
@@ -75,7 +75,7 @@ class S8co : public IoTItem {
     void doByInterval() {
       co2_measure();
       if (_s8_co2 < 15000)
-        setValue(_s8_co2);
+        setValue((float)_s8_co2);
       else
         SerialPrint("E", "Sensor S8_uart", "Error");
     }
